mock through first to last

+++ src/echonet/smartmeter/parseE0.go
@@ -1,8 +1,9 @@
 package smartmeter
 
 // EPC 0xE0 積算電力量をパースする。
-func (sm *ELSmartMeterParser) ParseE0DeltaDenryoku(data []byte) (int, error) {
-	ret := -1
+// TODO mock
+func (sm *ELSmartMeterParser) ParseE0DeltaDenryoku(data []byte) (float64, error) {
+	ret := float64(-1)
 
 	err := sm.checkPreCondition()
 	if err != nil {
